KBB-102 [Refactor] global response package 이동
