+++ src/main/java/org/bootstrap/apicomposer/domain/post/contoller/PostController.java
@@ -3,9 +3,7 @@ package org.bootstrap.apicomposer.domain.post.contoller;
 import lombok.RequiredArgsConstructor;
 import org.bootstrap.apicomposer.domain.post.service.PostService;
 import org.bootstrap.apicomposer.domain.post.type.CategoryType;
-import org.bootstrap.apicomposer.global.common.SuccessResponse;
-import org.bootstrap.apicomposer.global.webclient.response.ApiResponse;
-import org.springframework.http.ResponseEntity;
+import org.bootstrap.apicomposer.global.common.response.ApiResponse;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
