--- src/main/java/org/bootstrap/apicomposer/global/common/SuccessCode.java
@@ -1,24 +1,0 @@
-package org.bootstrap.apicomposer.global.common;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public enum SuccessCode {
-    /**
-     * 200 Ok
-     */
-    OK(HttpStatus.OK, "요청이 성공했습니다."),
-
-    /**
-     * 201 Created
-     */
-    CREATED(HttpStatus.CREATED, "요청이 성공했습니다.");
-
-    private final HttpStatus httpStatus;
-    private final String message;
-}
-
