--- src/main/java/org/bootstrap/apicomposer/global/common/SuccessResponse.java
@@ -1,30 +1,0 @@
-package org.bootstrap.apicomposer.global.common;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder
-@Getter
-public class SuccessResponse<T> {
-    private int status;
-    private String message;
-    private T data;
-
-    public static <T> ResponseEntity<SuccessResponse<?>> ok(T data) {
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(SuccessResponse.of(SuccessCode.OK, data));
-    }
-
-    public static <T> SuccessResponse<?> of(SuccessCode successCode, T data) {
-        return SuccessResponse.builder()
-                .status(successCode.getHttpStatus().value())
-                .message(successCode.getMessage())
-                .data(data)
-                .build();
-    }
-}
