+++ src/main/java/org/bootstrap/apicomposer/global/common/response/ApiResponse.java
@@ -1,4 +1,4 @@
-package org.bootstrap.apicomposer.global.webclient.response;
+package org.bootstrap.apicomposer.global.common.response;
 
 import lombok.Builder;
 import lombok.Getter;
